+++ eth_defi/hotwallet.py
@@ -1,6 +1,13 @@
-"""Utilities for managing hot wallets."""
+"""Utilities for managing hot wallets.
+
+- Create local wallets from a private key
+
+- Sign transactions in batches
+
+"""
 
 import logging
+import secrets
 from decimal import Decimal
 from typing import Optional, NamedTuple
 
@@ -58,7 +65,7 @@ class HotWallet:
 
     .. note ::
 
-        Not thread safe. Manages consumed nonce counter locally.
+        Not thread safe. This class manages consumed nonce counter locally.
 
     """
 
@@ -90,7 +97,30 @@ class HotWallet:
     def sign_transaction_with_new_nonce(self, tx: dict) -> SignedTransactionWithNonce:
         """Signs a transaction and allocates a nonce for it.
 
-        :param: Ethereum transaction data as a dict. This is modified in-place to include nonce.
+        Example:
+
+        .. code-block:: python
+
+            web3 = Web3(mev_blocker_provider)
+            wallet = HotWallet.create_for_testing(web3)
+
+            # Send some ETH to zero address from
+            # the hot wallet
+            signed_tx = wallet.sign_transaction_with_new_nonce({
+                "from": wallet.address,
+                "to": ZERO_ADDRESS,
+                "value": 1,
+                "gas": 100_000,
+                "gasPrice": web3.eth.gas_price,
+            })
+            tx_hash = web3.eth.send_raw_transaction(signed_tx.rawTransaction)
+
+        :param tx:
+            Ethereum transaction data as a dict.
+            This is modified in-place to include nonce.
+
+        :return:
+            A transaction payload and nonce with used to generate this transaction.
         """
         assert type(tx) == dict
         assert "nonce" not in tx
@@ -163,3 +193,42 @@ class HotWallet:
         assert key.startswith("0x")
         account = Account.from_key(key)
         return HotWallet(account)
+
+    @staticmethod
+    def create_for_testing(web3: Web3, test_account_n=0, eth_amount=10) -> "HotWallet":
+        """Creates a new hot wallet and seeds it with ETH from one of well-known test accounts.
+
+        Shortcut method for unit testing.
+
+        Example:
+
+        .. code-block:: python
+
+            web3 = Web3(test_provider)
+            wallet = HotWallet.create_for_testing(web3)
+
+            signed_tx = wallet.sign_transaction_with_new_nonce(
+                {
+                    "from": wallet.address,
+                    "to": ZERO_ADDRESS,
+                    "value": 1,
+                    "gas": 100_000,
+                    "gasPrice": web3.eth.gas_price,
+                }
+            )
+
+            tx_hash = web3.eth.send_raw_transaction(signed_tx.rawTransaction)
+            assert_transaction_success_with_explanation(web3, tx_hash)
+
+        """
+        wallet = HotWallet.from_private_key("0x" + secrets.token_hex(32))
+        tx_hash = web3.eth.send_transaction(
+            {
+                "from": web3.eth.accounts[test_account_n],
+                "to": wallet.address,
+                "value": eth_amount * 10**18,
+            }
+        )
+        web3.eth.wait_for_transaction_receipt(tx_hash)
+        wallet.sync_nonce(web3)
+        return wallet